remake logs

+++ microservices/cameras/internal/service/transport.go
@@ -9,23 +9,23 @@ import (
 	"io"
 	"net/http"
 
+	"github.com/Impisigmatus/service_core/log"
 	"github.com/Impisigmatus/service_core/utils"
 	"github.com/LeonKote/PSSVTelegramBot/microservices/cameras/autogen/server"
 	"github.com/LeonKote/PSSVTelegramBot/microservices/cameras/internal/models"
 	"github.com/google/uuid"
 	jsoniter "github.com/json-iterator/go"
 	"github.com/minio/minio-go/v7"
+	"github.com/rs/zerolog"
 )
 
 type Transport struct {
 	app *Application
-	ctx context.Context
 }
 
-func NewTransport(ctx context.Context, app *Application) server.ServerInterface {
+func NewTransport(app *Application) server.ServerInterface {
 	return &Transport{
 		app: app,
-		ctx: ctx,
 	}
 }
 
@@ -45,24 +45,30 @@ func NewTransport(ctx context.Context, app *Application) server.ServerInterface 
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) PostApiCamerasAdd(w http.ResponseWriter, r *http.Request) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	data, err := io.ReadAll(r.Body)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
 		return
 	}
 
 	var camera models.Camera
 	if err := jsoniter.Unmarshal(data, &camera); err != nil {
-		utils.WriteString(w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
+		utils.WriteString(log, w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
 		return
 	}
 
 	if err := transport.app.repo.AddCamera(camera); err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось добавить пользователя")
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось добавить пользователя")
 		return
 	}
 
-	utils.WriteNoContent(w)
+	utils.WriteNoContent(log, w)
 }
 
 // Set godoc
@@ -80,17 +86,23 @@ func (transport *Transport) PostApiCamerasAdd(w http.ResponseWriter, r *http.Req
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) DeleteApiCamerasDeleteName(w http.ResponseWriter, r *http.Request, name string) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	ok, err := transport.app.repo.RemoveCamera(name)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Пользователя не существует")
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Пользователя не существует")
 		return
 	}
 
 	if ok {
-		utils.WriteString(w, http.StatusOK, nil, "Пользователя не существует")
+		utils.WriteString(log, w, http.StatusOK, nil, "Пользователя не существует")
 		return
 	} else {
-		utils.WriteNoContent(w)
+		utils.WriteNoContent(log, w)
 		return
 	}
 }
@@ -111,18 +123,24 @@ func (transport *Transport) DeleteApiCamerasDeleteName(w http.ResponseWriter, r 
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiCamerasGetName(w http.ResponseWriter, r *http.Request, name string) {
-	user, err := transport.app.repo.GetCameraByName(name)
-	if err != nil {
-		if errors.Is(err, sql.ErrNoRows) {
-			utils.WriteNoContent(w)
-			return
-		}
-
-		utils.WriteString(w, http.StatusNoContent, err, "Не удалось получить данные о камере")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	utils.WriteObject(w, user)
+	user, err := transport.app.repo.GetCameraByName(name)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
+			utils.WriteNoContent(log, w)
+			return
+		}
+
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось получить данные о камере")
+		return
+	}
+
+	utils.WriteObject(log, w, user)
 }
 
 // Set godoc
@@ -139,17 +157,23 @@ func (transport *Transport) GetApiCamerasGetName(w http.ResponseWriter, r *http.
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiCamerasGet(w http.ResponseWriter, r *http.Request) {
-	cameras, err := transport.app.repo.GetAllCameras()
-	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось получить список камер")
-		return
-	}
-	if len(cameras) == 0 {
-		utils.WriteString(w, http.StatusInternalServerError, err, "В базе нет камер")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	utils.WriteObject(w, cameras)
+	cameras, err := transport.app.repo.GetAllCameras()
+	if err != nil {
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось получить список камер")
+		return
+	}
+	if len(cameras) == 0 {
+		utils.WriteString(log, w, http.StatusInternalServerError, err, "В базе нет камер")
+		return
+	}
+
+	utils.WriteObject(log, w, cameras)
 }
 
 // Set godoc
@@ -168,34 +192,41 @@ func (transport *Transport) GetApiCamerasGet(w http.ResponseWriter, r *http.Requ
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) PostApiCamerasRecord(w http.ResponseWriter, r *http.Request) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	reqId := uuid.New().String()
 
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
 		return
 	}
 
 	go func() {
+		ctx := context.Background()
 		var record models.Record
 		if err := jsoniter.Unmarshal(body, &record); err != nil {
-			utils.WriteString(w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
+			utils.WriteString(log, w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
 			return
 		}
 
-		fileName, err := transport.app.Record(transport.ctx, record, true, reqId)
+		fileName, err := transport.app.Record(log, ctx, record, true, reqId)
 		if err != nil {
-			if err := transport.app.ChangeStatus(fileName, nilFileSize, statusFailed); err != nil {
-				utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось добавить запись в очередь")
+			if err := transport.app.ChangeStatus(log, fileName, nilFileSize, statusFailed); err != nil {
+				utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось добавить запись в очередь")
 				return
 			}
 
-			utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось записать видео")
+			utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось записать видео")
 			return
 		}
 	}()
 
-	utils.WriteObject(w, models.Uuid{Uuid: reqId})
+	utils.WriteObject(log, w, models.Uuid{Uuid: reqId})
 }
 
 // Set godoc
@@ -214,35 +245,42 @@ func (transport *Transport) PostApiCamerasRecord(w http.ResponseWriter, r *http.
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) PostApiCamerasCapture(w http.ResponseWriter, r *http.Request) {
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
+		return
+	}
+
 	reqId := uuid.New().String()
-	utils.WriteObject(w, models.Uuid{Uuid: reqId})
+	utils.WriteObject(log, w, models.Uuid{Uuid: reqId})
 
 	body, err := io.ReadAll(r.Body)
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read body: %s", err), "Не удалось прочитать тело запроса")
 		return
 	}
 
 	go func() {
+		ctx := context.Background()
 		var record models.Record
 		if err := jsoniter.Unmarshal(body, &record); err != nil {
-			utils.WriteString(w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
+			utils.WriteString(log, w, http.StatusBadRequest, fmt.Errorf("Invalid parse body: %s", err), "Не удалось распарсить тело запроса формата JSON")
 			return
 		}
 
-		fileName, err := transport.app.Record(transport.ctx, record, false, reqId)
+		fileName, err := transport.app.Record(log, ctx, record, false, reqId)
 		if err != nil {
-			if err := transport.app.ChangeStatus(fileName, nilFileSize, statusFailed); err != nil {
-				utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось добавить запись в очередь")
+			if err := transport.app.ChangeStatus(log, fileName, nilFileSize, statusFailed); err != nil {
+				utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось добавить запись в очередь")
 				return
 			}
 
-			utils.WriteString(w, http.StatusInternalServerError, err, "Не удалось записать видео")
+			utils.WriteString(log, w, http.StatusInternalServerError, err, "Не удалось записать видео")
 			return
 		}
 	}()
 
-	utils.WriteNoContent(w)
+	utils.WriteNoContent(log, w)
 }
 
 // Set godoc
@@ -262,32 +300,38 @@ func (transport *Transport) PostApiCamerasCapture(w http.ResponseWriter, r *http
 // @Failure 401 {object} nil "Ошибка авторизации"
 // @Failure 500 {object} nil "Произошла внутренняя ошибка сервера"
 func (transport *Transport) GetApiCamerasDirFileNameGet(w http.ResponseWriter, r *http.Request, dir string, fileName string) {
-	path := fmt.Sprintf("/%s/%s", dir, fileName)
-	_, err := transport.app.minioClient.StatObject(transport.ctx, transport.app.bucketName, path, minio.StatObjectOptions{})
-	if err != nil {
-		if minio.ToErrorResponse(err).Code == "NoSuchKey" {
-			utils.WriteString(w, http.StatusNotFound, fmt.Errorf("File not found: %s", path), "Файл не найден")
-			return
-		}
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Error checking object existence: %v", err), "Ошибка проверки существования файла")
+	log, ok := r.Context().Value(log.CtxKey).(zerolog.Logger)
+	if !ok {
+		utils.WriteString(zerolog.Logger{}, w, http.StatusInternalServerError, fmt.Errorf("Invalid logger"), "Невалидный логгер")
 		return
 	}
 
-	object, err := transport.app.minioClient.GetObject(transport.ctx, transport.app.bucketName, path, minio.GetObjectOptions{})
+	path := fmt.Sprintf("/%s/%s", dir, fileName)
+	_, err := transport.app.minioClient.StatObject(r.Context(), transport.app.bucketName, path, minio.StatObjectOptions{})
 	if err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read object: %s", err), "Не удалось получить файл")
+		if minio.ToErrorResponse(err).Code == "NoSuchKey" {
+			utils.WriteString(log, w, http.StatusNotFound, fmt.Errorf("File not found: %s", path), "Файл не найден")
+			return
+		}
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Error checking object existence: %v", err), "Ошибка проверки существования файла")
+		return
+	}
+
+	object, err := transport.app.minioClient.GetObject(r.Context(), transport.app.bucketName, path, minio.GetObjectOptions{})
+	if err != nil {
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read object: %s", err), "Не удалось получить файл")
 		return
 	}
 	defer object.Close()
 
 	var buffer bytes.Buffer
 	if _, err := buffer.ReadFrom(object); err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid read object: %s", err), "Не удалось прочитать файл")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid read object: %s", err), "Не удалось прочитать файл")
 		return
 	}
 
 	if _, err := io.Copy(w, &buffer); err != nil {
-		utils.WriteString(w, http.StatusInternalServerError, fmt.Errorf("Invalid send object: %s", err), "Не удалось прочитать файл")
+		utils.WriteString(log, w, http.StatusInternalServerError, fmt.Errorf("Invalid send object: %s", err), "Не удалось прочитать файл")
 		return
 	}
 }
